Create artistList.html
Where I have stated the fake bands names would become class="bandName" and will be either in another link or pulled from php. whatever fits best. This is only a fundamental layout. And I hope it what is agreed on as I had my computer and files corrupt and required a reformat.